refactor: share real-time experiment plan
